# Conflicts:
#	helm/chouette/values.yaml

+++ mobi.chouette.exchange.neptune/src/main/java/mobi/chouette/exchange/neptune/parser/AccessLinkParser.java
@@ -22,9 +22,9 @@ import mobi.chouette.model.type.UserNeedEnum;
 import mobi.chouette.model.util.ObjectFactory;
 import mobi.chouette.model.util.Referential;
 
-import org.joda.time.Duration;
-import org.joda.time.LocalDateTime;
-import org.joda.time.LocalTime;
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
 import org.xmlpull.v1.XmlPullParser;
 
 @Log4j
@@ -51,7 +51,7 @@ public class AccessLinkParser implements Parser, Constant {
 				accessLink = ObjectFactory.getAccessLink(referential, objectId);
 				accessLink.setFilled(true);
 			} else if (xpp.getName().equals("objectVersion")) {
-				Integer version = ParserUtils.getInt(xpp.nextText());
+				Long version = ParserUtils.getLong(xpp.nextText());
 				accessLink.setObjectVersion(version);
 			} else if (xpp.getName().equals("creationTime")) {
 				LocalDateTime creationTime = ParserUtils.getLocalDateTime(xpp.nextText());
