+++ mobi.chouette.exchange.neptune/src/main/java/mobi/chouette/exchange/neptune/parser/ChouetteAreaParser.java
@@ -30,7 +30,7 @@ import mobi.chouette.model.util.Referential;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
-import org.joda.time.LocalDateTime;
+import java.time.LocalDateTime;
 import org.xmlpull.v1.XmlPullParser;
 
 @Log4j
@@ -84,7 +84,7 @@ public class ChouetteAreaParser implements Parser, Constant, JsonExtension {
 				stopArea = ObjectFactory.getStopArea(referential, objectId);
 				stopArea.setFilled(true);
 			} else if (xpp.getName().equals("objectVersion")) {
-				Integer version = ParserUtils.getInt(xpp.nextText());
+				Long version = ParserUtils.getLong(xpp.nextText());
 				stopArea.setObjectVersion(version);
 			} else if (xpp.getName().equals("creationTime")) {
 				LocalDateTime creationTime = ParserUtils.getLocalDateTime(xpp.nextText());
