+++ mobi.chouette.exchange.neptune/src/main/java/mobi/chouette/exchange/neptune/parser/ChouetteRouteParser.java
@@ -20,7 +20,7 @@ import mobi.chouette.model.type.PTDirectionEnum;
 import mobi.chouette.model.util.ObjectFactory;
 import mobi.chouette.model.util.Referential;
 
-import org.joda.time.LocalDateTime;
+import java.time.LocalDateTime;
 import org.xmlpull.v1.XmlPullParser;
 
 @Log4j
@@ -49,7 +49,7 @@ public class ChouetteRouteParser implements Parser, Constant {
 				route = ObjectFactory.getRoute(referential, objectId);
 				route.setFilled(true);
 			} else if (xpp.getName().equals("objectVersion")) {
-				Integer version = ParserUtils.getInt(xpp.nextText());
+				Long version = ParserUtils.getLong(xpp.nextText());
 				route.setObjectVersion(version);
 			} else if (xpp.getName().equals("creationTime")) {
 				LocalDateTime creationTime = ParserUtils.getLocalDateTime(xpp.nextText());
